# api/openapi.yaml
components:
  schemas:
...
    AgentView:
      additionalProperties: false
      properties:
        created_at:
          format: date-time
          type: string
        domain:
          type: string
        domain_verified:
          type: boolean
        email:
          type: string
        hitl_enabled:
          type: boolean
        hitl_expiration_action:
          type: string
        hitl_ttl_seconds:
          format: int64
          type: integer
        id:
          type: string
        name:
          type: string
      required:
        - id
        - domain
        - email
        - name
        - domain_verified
        - created_at
        - hitl_enabled
        - hitl_ttl_seconds
        - hitl_expiration_action
      type: object
...
    CreateAgentRequest:
      additionalProperties: false
      properties:
        email:
          type: string
        name:
          type: string
        slug:
          type: string
      type: object
    CreateAgentResponse:
      additionalProperties: false
      properties:
        domain:
          type: string
        email:
          type: string
        id:
          type: string
      required:
        - id
        - domain
        - email
      type: object
...
    DeleteAgentOutputBody:
      additionalProperties: false
      properties:
        status:
          type: string
      required:
        - status
      type: object
...
    ErrorEnvelope:
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
      required:
        - error
      type: object
...
    ListAgentsOutputBody:
      additionalProperties: false
      properties:
        agents:
          items:
            $ref: "#/components/schemas/AgentView"
          type:
            - array
            - "null"
      required:
        - agents
      type: object
...
    UpdateAgentRequest:
      additionalProperties: false
      properties:
        hitl_enabled:
          type: boolean
        hitl_expiration_action:
          type: string
        hitl_ttl_seconds:
          format: int64
          type: integer
      type: object
...
  securitySchemes:
    bearer:
      description: "API key (e2a_…) or OAuth 2.1 access token, sent as `Authorization: Bearer <token>`."
      scheme: bearer
      type: http
info:
  description: e2a — authenticated email gateway for AI agents. v1 contract.
  title: e2a API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/agents:
    get:
      description: List the agents owned by the authenticated account.
      operationId: listAgents
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAgentsOutputBody"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: List agents
      tags:
        - agents
    post:
      description: Register an agent on a verified domain the caller owns (or, when slug registration is enabled, on the shared domain).
      operationId: createAgent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAgentResponse"
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Create an agent
      tags:
        - agents
  /v1/agents/{address}:
    delete:
      description: Delete an agent the caller owns.
      operationId: deleteAgent
      parameters:
        - description: The agent's full email address, e.g. support@acme.com.
          in: path
          name: address
          required: true
          schema:
            description: The agent's full email address, e.g. support@acme.com.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAgentOutputBody"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Delete an agent
      tags:
        - agents
    get:
      description: Fetch a single agent the authenticated account owns, by full email address.
      operationId: getAgent
      parameters:
        - description: The agent's full email address, e.g. support@acme.com.
          in: path
          name: address
          required: true
          schema:
            description: The agent's full email address, e.g. support@acme.com.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Get an agent
      tags:
        - agents
    patch:
      description: Patch an agent's HITL settings. Returns the post-update agent.
      operationId: updateAgent
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Update an agent
      tags:
        - agents
  /v1/agents/{address}/conversations:
    get:
      description: List an agent's conversation threads (derived from messages.conversation_id).
      operationId: listConversations
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
        - description: RFC3339.
          explode: false
          in: query
          name: since
          schema:
            description: RFC3339.
            type: string
        - description: RFC3339.
          explode: false
          in: query
          name: until
          schema:
            description: RFC3339.
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 100
            format: int64
            maximum: 200
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageConversationSummaryView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: List conversations
      tags:
        - conversations
  /v1/agents/{address}/conversations/{id}:
    get:
      description: Fetch a single conversation thread with its participants, labels, and member messages.
      operationId: getConversation
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetailView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Get a conversation
      tags:
        - conversations
  /v1/agents/{address}/messages:
    get:
      description: List an agent's messages (inbound + outbound) with filters and cursor pagination. Held outbound drafts appear as status=pending_approval.
      operationId: listMessages
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
        - description: Defaults to inbound.
          explode: false
          in: query
          name: direction
          schema:
            description: Defaults to inbound.
            enum:
              - inbound
              - outbound
              - all
            type: string
        - description: Inbound only. Defaults to unread for inbound, all otherwise.
          explode: false
          in: query
          name: status
          schema:
            description: Inbound only. Defaults to unread for inbound, all otherwise.
            enum:
              - unread
              - read
              - all
            type: string
        - description: Defaults to desc (newest first).
          explode: false
          in: query
          name: sort
          schema:
            description: Defaults to desc (newest first).
            enum:
              - asc
              - desc
            type: string
        - description: Case-insensitive substring match on sender.
          explode: false
          in: query
          name: from
          schema:
            description: Case-insensitive substring match on sender.
            type: string
        - description: Case-insensitive substring match on subject.
          explode: false
          in: query
          name: subject_contains
          schema:
            description: Case-insensitive substring match on subject.
            type: string
        - explode: false
          in: query
          name: conversation_id
          schema:
            type: string
        - description: Repeatable; AND-matched.
          explode: false
          in: query
          name: labels
          schema:
            description: Repeatable; AND-matched.
            items:
              type: string
            type:
              - array
              - "null"
        - description: RFC3339; created_at >= since.
          explode: false
          in: query
          name: since
          schema:
            description: RFC3339; created_at >= since.
            type: string
        - description: RFC3339; created_at < until.
          explode: false
          in: query
          name: until
          schema:
            description: RFC3339; created_at < until.
            type: string
        - explode: false
          in: query
          name: cursor
          schema:
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 50
            format: int64
            maximum: 100
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageMessageSummaryView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: List messages
      tags:
        - messages
    post:
      description: Send a new email from the agent named in the path (a new thread). The sender is the path agent — `reply`/`forward` are their own sub-resources. 202 + pending_approval when the agent has HITL enabled. Honors Idempotency-Key.
      operationId: sendMessage
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
        - in: header
          name: Idempotency-Key
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailRequest"
          application/octet-stream:
            schema:
              contentMediaType: application/octet-stream
              format: binary
              type: string
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResultView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Send a new email
      tags:
        - messages
  /v1/agents/{address}/messages/{id}:
    get:
      description: Fetch a single message (inbound or outbound) by id, scoped to an agent the caller owns. Includes the raw message and inbound auth headers.
      operationId: getMessage
      parameters:
        - description: The agent's full email address.
          in: path
          name: address
          required: true
          schema:
            description: The agent's full email address.
            type: string
        - description: The message id, e.g. msg_abc123.
          in: path
          name: id
          required: true
          schema:
            description: The message id, e.g. msg_abc123.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Get a message
      tags:
        - messages
...
  /v1/domains:
    get:
      operationId: listDomains
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListDomainsOutputBody"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: List domains
      tags:
        - domains
    post:
      operationId: registerDomain
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDomainRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainView"
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Register a domain
      tags:
        - domains
  /v1/domains/{domain}:
    delete:
      operationId: deleteDomain
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Delete a domain
      tags:
        - domains
    get:
      operationId: getDomain
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Get a domain
      tags:
        - domains
    patch:
      operationId: updateDomain
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDomainRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Update a domain (set primary)
      tags:
        - domains
  /v1/domains/{domain}/verify:
    post:
      description: Probe the domain's published DNS and, when the verification TXT is present, mark it verified. Returns the per-record diagnostic; a missing TXT yields 412.
      operationId: verifyDomain
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyDomainView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Verify a domain
      tags:
        - domains
  /v1/events:
    get:
      description: The webhook-event delivery log, filterable by type/agent/conversation/message and time range, with cursor pagination.
      operationId: listEvents
      parameters:
        - explode: false
          in: query
          name: type
          schema:
            type: string
        - explode: false
          in: query
          name: agent_id
          schema:
            type: string
        - explode: false
          in: query
          name: conversation_id
          schema:
            type: string
        - explode: false
          in: query
          name: message_id
          schema:
            type: string
        - description: RFC3339.
          explode: false
          in: query
          name: since
          schema:
            description: RFC3339.
            type: string
        - description: RFC3339.
          explode: false
          in: query
          name: until
          schema:
            description: RFC3339.
            type: string
        - explode: false
          in: query
          name: cursor
          schema:
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 50
            format: int64
            maximum: 200
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageEventJSON"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: List events
      tags:
        - events
...
  /v1/webhooks:
    get:
      operationId: listWebhooks
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWebhooksOutputBody"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: List webhooks
      tags:
        - webhooks
    post:
      operationId: createWebhook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookView"
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - bearer: []
      summary: Create a webhook
      tags:
        - webhooks
...